+++ R/zzz_exports.R
@@ -1,0 +1,19 @@
+#' @importFrom magrittr %>%
+#' @export
+magrittr::`%>%`
+
+#' @importFrom haven as_factor
+#' @export
+haven::as_factor
+
+#' @importFrom sjmisc get_label
+#' @export
+sjmisc::get_label
+
+#' @importFrom sjmisc get_labels
+#' @export
+sjmisc::get_labels
+
+#' @importFrom sjmisc word_wrap
+#' @export
+sjmisc::word_wrap